Add files via upload
I made folder so as to seperate css from the html. In essence I am creating external css folder which will be called upon in the html file. This will make the code cleaner, as well as allow me to only have to change the css in that folder and it will change the code site wide, instead of me having to look through the massive site of code and change the css individually.